Batch processing and 32bit value_t (#36)
Implemented a batch processing approach for construct result in order to prevent iterating over columns
during the procedure, waiting till we have a full page worth of data to fill and building column by column not
row by row.

Implemented batch copies of dense pages in order to avoid iterating all values. Now pages are copied at once
if they are found to be dense. Same applies for reconstructing the final ColumnarTable.

value_t is now 32bit packing 17 bits for page index and 15 bits for offset in case of VARCHAR this might change
in order to allow more pages seeing as 15 bits for the offset index is unnecessary.

Also clang-format was run on all the source code.