Adding final implentation (#37)
Adding a sophisticated matchcollector in order to later build the table by columns. Integrated copy_scan and to_columnar,
within the join allowing the hashtable to be built from ColumnarTable and the probe phase to be done with a ColumnarTable allowing construct result to build the intermediate representation from both value_t and ColumnarTable.
The same applies for to_columnar at the root node.